requirements.txt
requests
openai
runwayml
python-dotenv
flask 
sqlalchemy
flask-sqlalchemy
